+++ libs/utils/util.js
@@ -1,0 +1,75 @@
+'use strict';
+
+// Copyright 2019 OpenST Ltd.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// ----------------------------------------------------------------------------
+//
+// http://www.simpletoken.org/
+//
+// ----------------------------------------------------------------------------
+
+const Web3Utils = require('web3-utils');
+const Crypto = require('crypto');
+
+/**
+ * This class includes the utitity functions.
+ * @class
+ * @classdesc Provides the common utility functions.
+ */
+class Util {
+  /**
+   * @typedef {Object} HashLock
+   *
+   * @property {string} secret The string that was used to generate hash lock
+   * and unlock secret.
+   *
+   * @property {string} unlockSecret The unlock secret string.
+   *
+   * @property {string} hashLock The hash lock.
+   */
+
+  /**
+   * @function createSecretHashLock
+   *
+   * Creates a random secret string, unlock secrete and hashlock.
+   *
+   * @returns {HashLock} HashLock object.
+   */
+  static createSecretHashLock() {
+    let secret = Crypto.randomBytes(16).toString('hex');
+    return Util.toHashLock(secret);
+  }
+
+  /**
+   * @function toHashLock
+   *
+   * Returns the HashLock from the given secret string.
+   *
+   * @param {string} secretString The secret string.
+   *
+   * @returns {HashLock} HashLock object.
+   */
+  static toHashLock(secretString) {
+    let secretBytes = Buffer.from(secretString);
+
+    return {
+      secret: secretString,
+      unlockSecret: `0x${secretBytes.toString('hex')}`,
+      hashLock: Web3Utils.keccak256(secretString)
+    };
+  }
+}
+
+module.exports = Util;
