+++ package.json
@@ -28,10 +28,12 @@
     "webpack-build": "webpack --mode=production",
     "web-pre-build": "node webPreBuild.js",
     "web-build": "npm-run-all web-pre-build webpack-build",
-    "web-post-build": "git add dist/."
+    "web-post-build": "git add dist/.",
+    "test": "mocha"
   },
   "dependencies": {
     "bignumber.js": "4.1.0",
+    "eslint-config-airbnb": "^17.1.0",
     "node-cmd": "3.0.0",
     "path": "0.12.7",
     "shelljs": "0.8.1",
@@ -46,18 +48,22 @@
     "chai": "4.2.0",
     "commander": "2.8.1",
     "edit-json-file": "1.0.8",
+    "eslint": "^5.12.1",
+    "eslint-config-airbnb-base": "^13.1.0",
+    "eslint-plugin-import": "^2.15.0",
     "expose-loader": "0.7.5",
     "lint-staged": "8.0.3",
     "mocha": "5.2.0",
     "npm-run-all": "4.1.3",
     "os": "0.1.1",
+    "package-json-cleanup-loader": "1.0.3",
     "pre-commit": "1.2.2",
     "prettier": "1.15.3",
+    "sinon": "7.2.3",
     "string-replace-loader": "2.1.1",
     "webpack": "4.19.1",
     "webpack-cli": "3.1.0",
-    "webpack-uglify-js-plugin": "1.1.9",
-    "package-json-cleanup-loader": "1.0.3"
+    "webpack-uglify-js-plugin": "1.1.9"
   },
   "pre-commit": {
     "run": [
